requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
opendartreader==0.2.3
supabase==2.21.1
requests==2.32.5
pandas>=0.19.2
python-dotenv==1.0.0
yfinance==0.2.28
psycopg2-binary==2.9.9
fredapi==0.5.2
